friend api 수정

+++ server/router/friends.js
@@ -9,30 +9,34 @@ module.exports = function(app, fs, db){
         else res.status(401).json({error:{status:401,desc:"401 error - Unauthorized"}});
     })
     // 자신의 대기중인 친구추가 요청리스트 보기
-    app.get('/friends/invite/:user_ID', async function(req, res){
+    app.get('/friends/invite', async function(req, res){
         if(req.session.passport){
-            var result = await db.getFreiendRequests(req.params.user_ID);
+            var user = req.session.passport.user;
+            var result = await db.getFreiendRequests(user.id);
             res.send(result);
         }
         else res.status(401).json({error:{status:401,desc:"401 error - Unauthorized"}});
     })
     // 자신의 친구 리스트 보기
-    app.get('/friends/:user_ID/list', async function(req, res){
+    app.get('/friends/list', async function(req, res){
         if(req.session.passport){
-            var result = await db.getFreiendList(req.params.user_ID);
+            // 유저 정보를 가져온다.
+            var user = req.session.passport.user;
+            var result = await db.getFreiendList(user.id);
             console.log(result);
             res.send(result);
         }
         else res.status(401).json({error:{status:401,desc:"401 error - Unauthorized"}});
     })
     // 친구 추가 요청
-    app.post('/friends/:friend_ID/request',async function(req, res){
+    app.post('/friends/:friend_ID',async function(req, res){
         if(req.session.passport){
             // 유저 정보를 가져온다.
             var user = req.session.passport.user;
             // 친구와의 관계를 가져온다.
-            var result = await db.checkFriend(user.id, req.params.friend_ID);
-            if((result.isFriend === false) && (result.isPending === false) ){
+            var result = await db.checkFriend(req.params.friend_ID, user.id);
+            var isUser = await db.isUser(req.params.friend_ID);
+            if((result.isFriend === false) && (result.isPending === false) && isUser ){
                 await db.addFriend(req.params.friend_ID, user.id);
                 res.status(202).json({response:{status:202,desc:"request success"}});
             }else{
@@ -42,13 +46,14 @@ module.exports = function(app, fs, db){
         else res.status(401).json({error:{status:401,desc:"401 error - Unauthorized"}});  
     })    
     // 친구추가 수락
-    app.post('/friends/:friend_ID',async function(req, res){
+    app.put('/friends/:friend_ID',async function(req, res){
         if(req.session.passport){
             // 유저 정보
             var user = req.session.passport.user;
             // 친구와의 관계 정보를 가져온다.
             var result = await db.checkFriend(user.id, req.params.friend_ID);
-            if((result.isFriend === false) && (result.isPending === true)){
+            var isUser = await db.isUser(req.params.friend_ID);
+            if((result.isFriend === false) && (result.isPending === true) && isUser){
                 await db.acceptFriend(user.id, req.params.friend_ID);
                 res.status(202).json({response:{status:202,desc:"request success"}});
             }else{
@@ -65,7 +70,8 @@ module.exports = function(app, fs, db){
             var user = req.session.passport.user;
             // 친구인지 정보를 가져온다.
             var result = await db.checkFriend(user.id, req.params.friend_ID);
-            if(result.isFriend === true){
+            var isUser = await db.isUser(req.params.friend_ID);
+            if(result.isFriend === true && isUser){
                 await db.deleteFriend(user.id, req.params.friend_ID);
                 res.status(202).json({response:{status:202,desc:"request success"}});
             }else {
